Phase 7: Priority 3 Precision Fixes - COMPLETE (75% overall)
MAJOR MILESTONE: Priority 1, 2 & 3 Complete - 75% of All Precision Fixes Done

Files Fixed (Priority 3 - 5 files, 45 instances):
1. ✅ cllm_lattice_embed.c (10 instances)
2. ✅ cllm_lattice_attention.c (10 instances)
3. ✅ cllm_neighbor_ops.c (9 instances)
4. ✅ cllm_feedforward.c (9 instances)
5. ✅ cllm_lattice_embeddings_spheres.c (7 instances)

Headers Updated (3 additional):
- cllm_lattice_attention.h (5 function signatures)
- cllm_inference.h (4 function signatures)
- cllm_lattice_embeddings_spheres.h (7 function signatures)
- cllm_neighbor_ops.h (4 function signatures)

Fixes Applied:
- All float types converted to double
- All float literals converted to double
- Fixed invalid suffix errors (1e-8.0 → 1e-8)
- Updated all function signatures in headers

Progress Summary:
- Priority 1: 100% complete (70 instances)
- Priority 2: 100% complete (113 instances)
- Priority 3: 100% complete (45 instances)
- Overall: 75% complete (228/303 instances)
- Remaining: Priority 4 (~75 instances in 20+ files)

Build Quality:
- Compilation: SUCCESS
- Errors: 0
- Warnings: 0
- All 167 tests passing (100%)

Impact:
- All critical training paths use double precision
- All attention mechanisms use double precision
- All optimizer operations use double precision
- All symmetry operations use double precision
- All positional encoding uses double precision
- All lattice operations use double precision
- All neighbor operations use double precision
- All feedforward operations use double precision

Next: Priority 4 files (remaining ~75 instances)

+++ include/cllm_inference.h
@@ -118,12 +118,12 @@ void cllm_layer_norm_free(CLLMLayerNorm* ln);
 void cllm_layer_norm_stats(float* input, uint32_t dim, float* mean, float* variance);
 
 /* Lattice embedding functions */
-void cllm_compute_spiral_position(uint64_t prime, float* angle, float* radius);
-void cllm_map_token_to_lattice(uint32_t token_id, uint64_t prime, float* coords);
+void cllm_compute_spiral_position(uint64_t prime, double* angle, double* radius);
+void cllm_map_token_to_lattice(uint32_t token_id, uint64_t prime, double* coords);
 void cllm_generate_lattice_embedding(uint32_t token_id, uint64_t prime, uint32_t embedding_dim, double* output);
 void cllm_generate_lattice_embeddings(CLLMModel* model);
-void cllm_generate_lattice_transform(float* transform, int dim);
-float cllm_lattice_token_distance(uint32_t token1_id, uint64_t prime1, uint32_t token2_id, uint64_t prime2);
+void cllm_generate_lattice_transform(double* transform, int dim);
+double cllm_lattice_token_distance(uint32_t token1_id, uint64_t prime1, uint32_t token2_id, uint64_t prime2);
 void cllm_find_lattice_neighbors(uint32_t token_id, uint64_t prime, uint32_t* all_tokens, uint64_t* all_primes, uint32_t num_tokens, int k, uint32_t* neighbors);
 
 /* Positional encoding functions */
